Fix: Skip Temporal Deployment and Cleanup in Preview Environment (#76)
* fix: skip deploying temporal-deployment.yaml in preview environment

* fix: skip Temporal cleanup in preview environment

* chore: add TODO with issue link for skipping temporal logic in preview

+++ lib/kube/clean.sh
@@ -30,10 +30,16 @@ fi
 
 if [ -d "$kube_env_dir" ]; then
     for file in "$kube_env_dir"/*; do
+        # TODO: https://github.com/jalantechnologies/github-ci/issues/77
+        if [[ "$KUBE_ENV" == "preview" && "$file" == *temporal-deployment.yaml ]]; then
+            echo "clean :: skipping temporal cleanup in preview env - $file"
+            continue
+        fi
         echo "clean :: cleaning from env config - $kube_env_dir/$file"
         envsubst <"$file" | kubectl delete --ignore-not-found=true -f -
     done
 fi
+
 
 # deployment post clean hook
 if [ -f "$kube_post_clean_script" ]; then
